refactor(#352,#353): split Tasks.cpp; extract displayTaskMain + mqttServiceTick stages
Tasks.cpp (1068) → Tasks.cpp core (IPC primitives, lifecycle, core-0
domain wrappers) + DisplayTask.cpp (core-1 display domain) +
ClockTask.cpp (1 Hz ticker), shared seams in TasksInternal.h (include
from the family's .cpp files ONLY — the WebEndpointsInternal.h rule).

displayTaskMain's 13-opcode switch is now pure dispatch into one static
exec* helper per opcode (#353); early case-breaks became returns.
mqttServiceTick split into mqttHandleOtaFreeze / mqttPublishEventDriven-
State / mqttPublishClusterSurfacing / mqttPublishTelemetry.

No logic change; 781 native + 77 pytest green.
